worked on arduino code a lil and convert ws payload to int[]

+++ controlling/src/main/java/me/autobot/lib/math/Mathf.java
@@ -69,7 +69,7 @@ public class Mathf {
         return Math.sin(r);
     }
 
-    public static int allPos(byte n) {
+    public static int allPos(int n) {
         if ((int) n < 0) return ((int) n) + 256;
         return n;
     }
